requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-generativeai==0.3.2
TTS
google-api-core
google-auth
google-cloud-core
google-cloud-storage
Werkzeug==3.1.3
Jinja2
requests
